My VA: removes EVSS claims code
resolve merge errors that removed necessary code
