+++ src/applications/personalization/common/mocks/users/index.js
@@ -10,6 +10,7 @@ const simpleUser = {
         'edu-benefits',
         'form-save-in-progress',
         'form-prefill',
+        'lighthouse',
         'evss-claims',
         'form526',
         'user-profile',
@@ -258,7 +259,7 @@ const cernerUser = {
         'edu-benefits',
         'form-save-in-progress',
         'form-prefill',
-        'evss-claims',
+        'lighthouse-claims',
         'form526',
         'user-profile',
         'appeals-status',
