+++ src/applications/personalization/dashboard/components/claims-and-appeals/HighlightedClaimAppeal.jsx
@@ -5,26 +5,19 @@ import { appealTypes } from '../../utils/appeals-v2-helpers';
 import Claim from './Claim';
 import Appeal from './Appeal';
 
-const HighlightedClaimAppeal = ({
-  claimOrAppeal,
-  name,
-  useLighthouseClaims = false,
-}) => {
+const HighlightedClaimAppeal = ({ claimOrAppeal, name }) => {
   if (!claimOrAppeal) {
     return <p>You have no claims or appeals updates in the last 30 days.</p>;
   }
   if (appealTypes.includes(claimOrAppeal.type)) {
     return <Appeal appeal={claimOrAppeal} name={name} />;
   }
-  return (
-    <Claim claim={claimOrAppeal} useLighthouseClaims={useLighthouseClaims} />
-  );
+  return <Claim claim={claimOrAppeal} />;
 };
 
 HighlightedClaimAppeal.propTypes = {
   claimOrAppeal: PropTypes.object.isRequired,
   name: PropTypes.string,
-  useLighthouseClaims: PropTypes.bool,
 };
 
 export default HighlightedClaimAppeal;
