+++ src/applications/personalization/dashboard/tests/e2e/appointments-errors.cypress.spec.js
@@ -21,7 +21,7 @@ describe('MyVA Dashboard - Appointments Error States', () => {
       '/v0/disability_compensation_form/rating_info',
       disabilityRating,
     );
-    cy.intercept('/v0/evss_claims_async', claimsSuccess());
+    cy.intercept('/v0/benefits_claims', claimsSuccess());
     cy.intercept('/v0/appeals', appealsSuccess());
     cy.intercept(
       '/v0/disability_compensation_form/rating_info',
