+++ src/applications/personalization/dashboard/tests/e2e/benefit-payments.cypress.spec.js
@@ -31,7 +31,7 @@ describe('The My VA Dashboard - Benefit Payments', () => {
     cy.login(mockUser);
     cy.intercept('/v0/profile/service_history', serviceHistory);
     cy.intercept('/v0/profile/full_name', fullName);
-    cy.intercept('/v0/evss_claims_async', claimsSuccess());
+    cy.intercept('/v0/benefits_claims', claimsSuccess());
     cy.intercept('/v0/appeals', appealsSuccess());
     cy.intercept(
       '/v0/disability_compensation_form/rating_info',
@@ -163,7 +163,7 @@ describe('when the payment history claims does not exist', () => {
     cy.login(mockUser1);
     cy.intercept('/v0/profile/service_history', serviceHistory);
     cy.intercept('/v0/profile/full_name', fullName);
-    cy.intercept('/v0/evss_claims_async', claimsSuccess());
+    cy.intercept('/v0/benefits_claims', claimsSuccess());
     cy.intercept('/v0/appeals', appealsSuccess());
     cy.intercept(
       '/v0/disability_compensation_form/rating_info',
@@ -227,7 +227,7 @@ describe('when the payment history claims is false', () => {
     cy.login(mockUser2);
     cy.intercept('/v0/profile/service_history', serviceHistory);
     cy.intercept('/v0/profile/full_name', fullName);
-    cy.intercept('/v0/evss_claims_async', claimsSuccess());
+    cy.intercept('/v0/benefits_claims', claimsSuccess());
     cy.intercept('/v0/appeals', appealsSuccess());
     cy.intercept(
       '/v0/disability_compensation_form/rating_info',
