+++ src/applications/personalization/profile/tests/fixtures/claims-success.js
@@ -5,72 +5,62 @@ const claimsSuccess = (updatedDaysAgo = 1, open = true) => {
     data: [
       {
         id: '266374',
-        type: 'evss_claims',
+        type: 'claim',
         attributes: {
-          evssId: 266374,
-          dateFiled: '2013-10-18',
-          minEstDate: '2014-06-03',
-          maxEstDate: '2014-06-08',
-          phaseChangeDate: daysAgo(updatedDaysAgo),
-          open,
-          waiverSubmitted: false,
-          documentsNeeded: false,
-          developmentLetterSent: true,
-          decisionLetterSent: false,
-          phase: open ? 2 : 8,
-          everPhaseBack: false,
-          currentPhaseBack: false,
-          requestedDecision: false,
+          claimDate: '2013-10-18',
+          claimPhaseDates: {
+            phaseChangeDate: daysAgo(updatedDaysAgo),
+          },
           claimType: 'Dependency',
-          updatedAt: '2021-01-07T20:47:32.248Z',
+          closeDate: open ? null : '2013-10-31',
+          decisionLetterSent: false,
+          developmentLetterSent: false,
+          documentsNeeded: false,
+          endProductCode: '404',
+          evidenceWaiverSubmitted5103: false,
+          lighthouseId: 266374,
+          status: open ? 'INITIAL_REVIEW' : 'COMPLETE',
         },
       },
       {
         id: '600121251',
-        type: 'evss_claims',
+        type: 'claim',
         attributes: {
-          evssId: 600121251,
-          dateFiled: '2018-01-10',
-          minEstDate: null,
-          maxEstDate: null,
-          phaseChangeDate: '2018-06-12',
-          open: false,
-          waiverSubmitted: false,
-          documentsNeeded: false,
-          developmentLetterSent: false,
-          decisionLetterSent: false,
-          phase: 8,
-          everPhaseBack: false,
-          currentPhaseBack: false,
-          requestedDecision: false,
+          claimDate: '2018-01-10',
+          claimPhaseDates: {
+            phaseChangeDate: '2018-06-12',
+          },
           claimType: 'Dependency',
-          updatedAt: '2021-02-08T16:51:48.335Z',
+          closeDate: daysAgo(90),
+          decisionLetterSent: false,
+          developmentLetterSent: false,
+          documentsNeeded: false,
+          endProductCode: '404',
+          evidenceWaiverSubmitted5103: false,
+          lighthouseId: 600121251,
+          status: 'COMPLETE',
         },
       },
       {
         id: '600090417',
-        type: 'evss_claims',
+        type: 'claim',
         attributes: {
-          evssId: 600090417,
-          dateFiled: '2017-01-02',
-          minEstDate: null,
-          maxEstDate: null,
-          phaseChangeDate: '2016-12-06',
-          open: false,
-          waiverSubmitted: false,
-          documentsNeeded: false,
-          developmentLetterSent: false,
-          decisionLetterSent: false,
-          phase: 8,
-          everPhaseBack: false,
-          currentPhaseBack: false,
-          requestedDecision: false,
+          claimDate: '2017-01-02',
+          claimPhaseDates: {
+            phaseChangeDate: '2016-12-06',
+          },
           claimType: 'Compensation',
-          updatedAt: '2021-02-16T21:17:41.458Z',
+          closeDate: daysAgo(90),
+          decisionLetterSent: false,
+          developmentLetterSent: false,
+          documentsNeeded: false,
+          endProductCode: '404',
+          evidenceWaiverSubmitted5103: false,
+          lighthouseId: 600090417,
+          status: 'COMPLETE',
         },
       },
     ],
-    meta: { syncStatus: 'SUCCESS' },
   };
 };
